Add Apex command to fetch current Apex Legends map
Introduces a new Apex command that retrieves and displays the current and next Apex Legends map using data from apexlegendsstatus.com. The command responds with an embed containing map details and an image.
